+++ src/Solvers/Solvers.jl
@@ -1,0 +1,9 @@
+module Solvers
+
+using Oceananigans.Grids
+
+abstract type PoissonBCs end
+
+include("utils.jl")
+
+end
